Updated templates, forms, models and views.
Set forms to save the user attribute to account models.

Added a __str__ method to account model to display the username of the referenced user, when viewed from the admin.

Passed request.user into user element when custom forms are called in views.

Added humanize to settings and bio template (not using it yet though).

Updated readme document.